% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frmselect.R
\name{frmselect}
\alias{frmselect}
\title{fractional model selection function based on quasi-likelihood}
\usage{
frmselect(x, y, criterion="AIC",linkfrac="logit", method="forward")
}
\arguments{
\item{x}{a numeric matrix, with column names, containing the values of the covariates.}

\item{y}{a numeric vector containing the values of the response variable. It should be between 0 and 1.}

\item{criterion}{model selection critetion. Available options: AIC, BIC, HQ.The default value is AIC.}

\item{linkfrac}{link function, Available options: logit, probit, loglog, cloglog, cauchit.The default value is logit.}

\item{method}{the mode of stepwise search and allsubsets. Available options: forward, backward, both, allsubsets.The default value is forward.}
}
\value{
A list contains information criterion, link function, model selection method,
minimal value of information criterion, the order of variables which are chosen in the model,
the names of corresonding variables and the estimated coefficients of them.
}
\description{
frmselect is used to select a formula-based model by information criterions of fractional regression models.
}
\examples{
N <- 250
u <- rnorm(N)
X <- cbind(rnorm(N),rnorm(N), rnorm(N),rnorm(N))
dimnames(X)[[2]] <- c("X1","X2","X3", "X4")
ym <- exp(X[,1]+X[,2]+X[,3]+X[,4]+u)/(1+exp(X[,1]+X[,2]+X[,3]+X[,4]+u))
y <- rbeta(N,ym*20,20*(1-ym))
y[y > 0.9] <- 1
frmselect(X,y, method = "forward")
}
