% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoselect.R
\name{lasso.select}
\alias{lasso.select}
\title{Lasso procedure based on bamlss}
\usage{
lasso.select(x, coefficient, threshold)
}
\arguments{
\item{x}{The independent variables}

\item{coefficent}{The coefficients estimated by bamlss()}

\item{threshold}{The threshold controls which coefficients should be set to 0. Default value is 1e-3. Coefficients smaller than threshold would be set to 0.}
}
\value{
A list of results that contains lambda parameters of common and factor variables,
lasso index which represents the index of coefficients that are set to be zero and the simplified coefficients.
}
\description{
This function is used to simplify the model based on lasso method with bamlss() function.
}
