% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaselect.R
\name{betaselect}
\alias{betaselect}
\title{fractional model selection function based on beta regression}
\usage{
betaselect(x, y, criterion, link, method)
}
\arguments{
\item{x}{a numeric matrix or data frame, with column names, containing the values of the covariates.}

\item{y}{a numeric vector containing the values of the response variable. It should be between 0 and 1.}

\item{criterion}{model selection critetion. Available options: AIC, BIC, HQ.The default value is AIC.}

\item{link}{link function, Available options: logit, probit, loglog, cloglog, log, cauchit.The default value is logit.}

\item{method}{the mode of stepwise search and allsubsets. Available options: forward, backward, both, allsubsets.The default value is forward.}
}
\value{
A list contains information criterion, link function, model selection method,
minimal value of information criterion, the order of variables which are chosen in the model,
the names of corresponding variables and the estimated coefficients of them.
}
\description{
betaselect is used to select a formula-based model by information criterions of fractional regression models.
}
\examples{
set.seed(123)
library(bamlss)
d <- GAMart()
x <- data.frame(d$x1, d$x2, d$x3)
y <- d$bnum
betaselect(x, y)
}
