% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamlss_link.R
\name{frm_bamlss}
\alias{frm_bamlss}
\title{Bamlss quasi-binomial family}
\usage{
frm_bamlss(link = "logit",...)
}
\arguments{
\item{link}{link function, Available options: logit, probit}

\item{...}{Arguments to pass to bamlss}
}
\description{
This is a quasi-binomial family which is suitable for bamlss() function in bamlss package.
This quasi-binomial family only contains two link functions, which are used mostly in the economics,
logit and probit function.
}
\details{
It is a family of bamlss package, can be used in the bamlss() function. The data type is
the fractional data which is between 0 and 1. Especially for probit link, the response data should be
strictly smaller than 1, otherwise it might be a warning that the backfitting algorithm did not converge!
}
\examples{
set.seed(123)
library(bamlss)
d <- GAMart()
b <- bamlss(bnum ~ x1 + x2 + x3, data = d, family = frm_bamlss(link = "logit") ,sampler = FALSE, multiple = FALSE)
}
